+++ apps/dsp-app/src/app/main/directive/input-masks.ts
@@ -1,0 +1,11 @@
+export class InputMasks {
+  public static wholeNumberInputMask(numbersCount: number) {
+    return {
+      mask: Array.from(new Array(numbersCount), () => new RegExp('\\d')),
+      showMask: false,
+      guide: true,
+      placeholderChar: ' ',
+      keepCharPositions: true,
+    };
+  }
+}
